companion 으로 변경
